% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{has_any_typst}
\alias{has_any_typst}
\alias{has_typst_cli}
\alias{has_quarto_typst}
\alias{has_quarto_cli}
\alias{typr_has_typst}
\alias{typr_has_quarto}
\alias{typr_has_quarto_typst}
\title{Find if Typst is available}
\usage{
has_any_typst()

has_typst_cli()

has_quarto_typst()

has_quarto_cli()

typr_has_typst()

typr_has_quarto()

typr_has_quarto_typst()
}
\value{
a logical if Typst or Quarto are found
}
\description{
Find if Typst is available
}
\examples{
has_any_typst()
has_typst_cli()
has_quarto_cli()
has_quarto_typst()
}
