% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{typr_info}
\alias{typr_info}
\title{Display Typst Debugging Information}
\usage{
typr_info()
}
\value{
a character vector of debugging information
}
\description{
Display Typst Debugging Information
}
\examples{
\dontshow{if (has_any_typst()) withAutoprint(\{ # examplesIf}
typr_info()
\dontshow{\}) # examplesIf}
}
