% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{typr_version}
\alias{typr_version}
\title{Get installed Typst version}
\usage{
typr_version(keep_hash = FALSE)
}
\arguments{
\item{keep_hash}{description}
}
\value{
a version, numeric if \code{!keep_hash}
}
\description{
Get installed Typst version
}
\examples{
\dontshow{if (has_any_typst()) withAutoprint(\{ # examplesIf}
typr_version()
\dontshow{\}) # examplesIf}
}
