% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions.R
\name{typr_completions}
\alias{typr_completions}
\title{Retrieve shell completions}
\usage{
typr_completions(shell = c("bash", "elvish", "fish", "powershell", "zsh"))
}
\arguments{
\item{shell}{The shell to generate completions for. One of
\code{c('bash', 'elvish', 'fish', 'powershell', 'zsh')}.}
}
\value{
a character vector of debugging information
}
\description{
Support for this function was added in Typst 0.14.0. If an earlier version
is in used, an empty character vector is returned.
}
\examples{
\dontshow{if (has_any_typst()) withAutoprint(\{ # examplesIf}
typr_completions(shell = 'powershell')
\dontshow{\}) # examplesIf}
}
