% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{typr_fonts}
\alias{typr_fonts}
\title{List Available Typst Fonts}
\usage{
typr_fonts()
}
\value{
a character vector of font names
}
\description{
List Available Typst Fonts
}
\examples{
\dontshow{if (has_any_typst()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
typr_fonts()
\dontshow{\}) # examplesIf}
}
