% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{typr_help}
\alias{typr_help}
\title{Get Help Information for a Typst command}
\usage{
typr_help(command = NULL)
}
\arguments{
\item{command}{name of a command to get help for, e.g. \code{'compile'}}
}
\value{
a character vector of information, invisibly
}
\description{
Prints a message to the terminal with help information.
Silently returns the content as character which can be printed with \code{cat()}.
}
\examples{
\dontshow{if (has_any_typst()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
typr_help('compile')
\dontshow{\}) # examplesIf}
}
