% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{typr_init}
\alias{typr_init}
\title{Create a project from a template}
\usage{
typr_init(template = NULL, version = NULL, dir = NULL, typst_args = NULL)
}
\arguments{
\item{template}{a template to use, can be local or on Typst Universe. If the
string does not specify \verb{@preview} or \verb{@local}, \code{'@preview'} will be prepended.}

\item{version}{a version string, e.g. \verb{0.1.0}}

\item{dir}{The project directory}

\item{typst_args}{Additional arguments to pass to Typst. Can be listed with \code{typr_help('init')}}
}
\value{
a directory, invisibly
}
\description{
Create a project from a template
}
\examples{
\dontshow{if (has_any_typst()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
typr_init('charged-ieee', dir = paste0(tempdir(), '/typr_example'))
\dontshow{\}) # examplesIf}
}
